Removes old Dewy application in favour of Node webapp
The old Dewy application was designed to work locally on a Drupal
server without any communication to a Dewy server and API. It required
setting up a Python and CouchDB environment - a very complicated
process that users won’t enjoy and many won’t be able to do.

This switches us to a new model where a Drupal module is all that is
needed to enable Dewy. This means setup for our users will take seconds
and work every time.