Creates backbone of what will be the Angular app
‘Backbone of the Angular app’ Get it? Okay, I’m not funny. But this is
the first real effort to moving the logic of page generation from
server-side Express to the client-side Angular app. This sets up nested
controllers and templates to drive the experience.

A lot needs to be done to move the Express app into more of an API
layer for the Angular app, but I can now see how it may take shape.