Improves add/remove rule buttons to work properly in all scenarios
When removing the last rule from a rule group, the rule group is
removed to.
When removing a rule group, any rules underneath it come up a level.
When there’s only one rule left in the entire filter, you can no longer
delete it.
When you add rules from a rule group it now goes into the rule group
instead of after it.