Refactors overview controllers and route resolving
The overview controllers (sites/modules/users/content) now get their
own route. This allows for resolving of what those controllers need
before the page is loaded, resulting in no flicker in the viewPage
area. This is less hacky, the menu has been updated everywhere, and the
controllers are in place. Just need to add logic and templating for
those new views.