Adds ability to use refresh token to get new access token on 401
This is done in the proxy API and then the new JWT is passed back to
Angular. It’s kind of sloppy though, since multiple requests can go to
the proxy API at once, resulting in multiple new access tokens getting
created. I’ll commit this for now, but I think I need to make the
Angular ask to get the new access token, not the proxy API.