Adds front end validation to Stripe.js subscription method
Also adds a setup for the Stripe Checkout method. The only issue with
that is that it does a post that doesn’t go through angular app and get
endowed with session information (JWT headers). Not sure how to get
around that yet. Whichever solution I can finish first wins!